renamed public facing things around big switch
when these functionalities were coded over 6 months ago, it was assumed the big switch (not then named) and the munnel opening would be simultaneous. It has been decided to just append the summer start to the "pre-munnel" functionality, and thus the current names are confusing. the file names will not be altered but the names the public sees will be